Massive improvements.
The horrid code for the scoring is now replaced with something simpler. Fully functional.
Still need to comment and do clean up,
In short: Sum up the rating a user has of books in their genre storing it in a list.
Do this for the potential match. Afterwards subtract the "opposite" genres.
Finally normalize the difference and sum that.
Giving you - = high match + = low match.

+++ app/src/main/java/com/example/shelfship/models/MatchesData.kt
@@ -12,5 +12,5 @@ data class Match(
 data class MatchScore(
     val uid: String = "",
     val username: String = "",
-    val score: Double = 0.0
+    val score: Float = 0f
 )
